+++ sentry-core/src/lib.rs
@@ -117,6 +117,7 @@ mod integration;
 mod intodsn;
 mod performance;
 mod scope;
+mod stream;
 mod transport;
 
 // public api or exports from this crate
@@ -132,6 +133,7 @@ pub use crate::integration::Integration;
 pub use crate::intodsn::IntoDsn;
 pub use crate::performance::*;
 pub use crate::scope::{Scope, ScopeGuard};
+pub use crate::stream::{SentryStream, SentryStreamExt};
 pub use crate::transport::{Transport, TransportFactory, TransportOptions};
 #[cfg(feature = "logs")]
 mod logger; // structured logging macros exported with `#[macro_export]`